druid: Bring druid1010.eqiad.wmnet into service
This host has been in setup, but is now being brought into service in
the druid_public cluster. This the second bullseye host in the cluster.
We are also adding the host to druid_public_hosts firewall block to allow
connection to the zookeper cluster required for the services
druid-middlemanager.service and druid-middlemanager.service.
This change also adds druid1009 to the druid_public_hosts firewall block

Bug: T336042
Change-Id: I76351cc37404af84e5298f09794eff3a153e3da3