dnsbox: add ntp.anycast.wmnet as the anycasted NTP address
This commit adds support for setting up the anycasted service that will
power ntp.anycast.wmnet. We announce this IP from each of the DNS boxes,
similar to the recdns anycast IP.

For a more detailed description of why we are doing this, please see the
task linked below. The TL;DR is that to reduce toil and manual updates
of the existing ntp.$site.wikimedia.org CNAME that points to a given DNS
host, we instead use an anycast address so that its consumers can use
one address everywhere and connect to the one nearest to them. In this
particular case, there is only one consumer: the d-i
clock-setup/ntp-server.

Hosts: dns1004.wikimedia.org, dns2004.wikimedia.org
Bug: T347054
Change-Id: I0cadd24bf8ec759ba31e482de79a69b94b860af9