ntp: lower the warning and critical thresholds for check_ntp_peer
The current warning and critical thresholds for check_ntp_peer should be
lowered as we are much closer to those values in practise and thus the
warning/critical thresholds should reflect that.

This was investigated and verified by Brandon by checking ntpq -pn
output via cumin.

Hosts: dns1004.wikimedia.org
Change-Id: I80d0c5fe61f88ee46f9a83fab69457049b204931