Support multiple spark yarn shufflers in parallel
This commit enables the use of two new hiera parameters within the
hadoop_clusters structure. The first is a boolean that enables the use
of multiple spark shufflers. The second is a hash of versions of spark
(major.minor) that will be enabled, along with the respective port
number on which they will listen.

When multiple shufflers are enabled at the cluster level, the jar files
implementing the service will be installed via debian packages (e.g.
spark-3.1-yarn-shuffler) as opposed to using a version that is shipped
by conda-analytics as we do at present.

This CR is intended to be a no-op as the default use of multiple
shufflers is set to false and has not yet been enabled for either
cluster.

Hosts: auto
Bug: T344910
Change-Id: I2ee904598882eb80bc2d06280ce9503af68f9c0d