cloudgw: add an IPv4 address for each node in the cloudgw <-> neutron subnet
Previous to this patch, the subnet for cloudgw <-> neutron transport was
a /30. Now it is a /29 so we can have an IPv4 address for each cloudgw node.

While at it, drop the `onlink` settings in the cloud realm routes, no longer
required after each node has an address in the subnet.
Same for the routing rule to select source address.

Bug: T347469
Bug: T348140
Hosts: re:cloudgw.*
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ib4c01b62be7f92130e3fc186369c47f6262ace6a