Deploy multiple spark shuffler services to the test cluster
We have previously enabled the parallel running of multiple versions of
the spark shuffler for yarn. This commit switches on that configuration
for the analytics-test-hadoop cluster.

In addition we configure the spark3 profile not to manage the symlink to
the version of the spark 3.1.2 version of the shuffler that is shipped
by conda-analytics. This will be handled by the spark-3.1-yarn-shuffler
debian package instead.

Hosts: P:hadoop::common,P:hadoop::worker
Bug: T344910
Change-Id: I7606f2f9523c49e69d9f03ec4e06e8969a2692e1