openldap: cross-validate-accounts wording
Why:
The reported message of:
  $name in nda group, but registered with a WMF account
can be confusing as is not clear based on what the script reports this a
WMF accont and if the data checked is just the LDAP one or also
data.yaml.

What:
Improve the wording by explicitly mentioning the email address domain
and the data.yaml file.

Change-Id: I7903e1cf6bc7015fe25907c5fe493df70935ef17
Hosts: auto