Support configuring the spark3 defaults with the default shuffler
We have enabled our YARN nodemanagers to run multiple versions of the
spark shuffler in parallel. The default shuffler in production is
currently version 3.1.2 and it is using the deault port number of 7001.

This commit enables us to update the spark3-defaults.conf file
cluster-wide to match a newly defined default setting. These settings
are configured to match the settings within the XML fragments that
configure the nodemanagers themselves.

Hosts: auto
Bug: T344910
Change-Id: I2f62c0f7a0f7b9c000f410153f87f85376ab07d7