Deploy multiple spark shufflers for yarn to production
We have created a means of running multiple spark shuffler services on
yarn. This commit deploys that configuration to production.

In addition we configure the spark3 profile not to manage the symlink to
the spark 3.1 shuffler that is shipped within conda-analytics. This
symlink will be managed by the spark-3.1-yarn-shuffler debian package
instead.

The existing symlinks (/usr/lib/hadoop-yarn/lib/spark3-yarn-shuffle.jar)
will be removed manually upon deployment, prior to a restart of the
nodemanager processes.

Hosts: P:hadoop::common
Bug: T344910
Change-Id: Ie07a0a6ca92e4cd619cdebf78d93e0fdaf082e58