parsoid-rt-client: Increase worker pool to 20 clients
With 16 clients, the test runs had the rare mysql crashers.
In addtiona, there have been a bunch of tweaks to the testreduce
server code which should reduce memory usage on the server.
Since we also have automatic restarts on crashes, we can try
to be a bit more aggressive and increase # of workers so the
test runs finish a bit faster.

Bug: T345220
Change-Id: I54929f7e82285668a309c3c4b932dfd85ed15239