late_command: update puppet installation logic
Currently the puppet5 backport is in the bookworm-wikimedia main repo as
such it is installed by default.

I have also updated this package to have a dependency on ruby-sorted-set
so there is no need to do anything for the default case.

when puppet 7 is selected for bookworm we need to increase the priority
of puppet from the debian repos.  with bullseye we need to add the
puppet7 component

Bug: T348319
Change-Id: If6bb37d9767e47fd3ece0ec711c835250c4101bc