hiera: announce ns1 IP from bird (codfw)
This commit announces the ns1 IP from bird for the codfw DNS hosts.
Currently, this is handled by static routes but as documented in the
task below, we want to offset this to bird.

Since the IP in question is an unicast address, we only announce it from
the given site, which is codfw for ns1. We achieve this by setting the
merge strategy in `P:bird::anycast', so this should be a NOOP for all
other (non-codfw) hosts.

Hosts: dns1004.wikimedia.org, dns2004.wikimedia.org, dns3003.wikimedia.org
Bug: T348041
Change-Id: If033a6d7a1a840edba00a6e14a8e4b47d409a584