role::redis::misc::{master,slave}: remove ORES configs
The ML team doesn't need anymore the ORES Redis instances
on rdb nodes (cache and queue).
This change removes the special configs for Pair 1 and 2
related to the ORES queue, to be able to re-use the instances
for other purposes.

The idea is to remove special configs for the ORES instances,
flush all data on them and mark the fact that they are unallocated
in https://wikitech.wikimedia.org/wiki/Redis#Cluster_redis_misc

Bug: T347278
Change-Id: I7e03946996e2958307b7613bc8e6b429f6531bb8