hiera: announce ns0 IP from bird (eqiad)
This commit announces the ns0 IP from bird for the eqiad DNS hosts.
Currently, this is handled by static routes but as documented in the
task below, we want to offset this to bird.

This is similar to the ec1752b10 but for eqiad instead.

Hosts: dns2004.wikimedia.org, dns1004.wikimedia.org, cloudservices1005.eqiad.wmnet
Bug: T348041
Change-Id: I2bed0ad7a2f93c8a00f1f03e8ad0bb6633e04ce4