Switch druid1006 zookeeper node with druid1011
We are onboarding new servers to the druid cluster druid10[09-11] the new servers are to replace
the soon to be decomissioned druid100[4-6] which are part of the ZooKeper cluster for
Druid public-eqiad cluster, (for AQS, wikistats, etc.) colocated on druid hosts.
We need to migrate the ZooKeeper service, one node at a time.

Bug: T336042
Change-Id: I4b066e089da997a3877859c7b47241bbb8008df0