Change eqiad cloudgw VIPs to /29 so system can ARP from them
Cloudgw's are currently sourcing all ARPs on the
cloud-instance-transport vlans from their unicast IP, as the VIP
they have on that interface has a /32 netmask.  Instead the
netmask should be set to /29, which is the updated size of
the subnet following T348140.

This is leading to occasional issues as cloudnet/neutron subnet
is configured with a /30 mask still (needs to change otherwise)
but either way we should have all IPs on the vlan configured
with matching subnet, so this changes cloudgw VIPs as needed.

Bug: T348140
Change-Id: Ibc8a1f3bc62fd4bb732716dabb0ffd73fa5ae58f