Ensure that alluxio cache directories are present for presto
We have previously worked with an alluxio cache directory on
/tmp/alluxio-cache where the presto worker already had the rights to
create it.

For our production cluster we will wish this to be under `/srv/cache` or
similar, so this commit allows the presto::catalog defiend type to
create its own cache directory and set the permissions accordingly.

We also set a systemd override so that when the cache files are created
in this directory they are not world-readable.

Hosts: auto
Bug: T266641
Change-Id: I9bf7a37d94437c525f3cd34c1b13c702b4af84f3