P:openstack::pdns::auth: make pdns web server listen on all IPs
Prometheus does not and will not have cloud-private access, but we still
want to monitor the pdns-auth service from the prod realm. To fix that,
this patch
* makes the pdns web server listen on all interfaces by setting the
  listen-on address as 0.0.0.0. This was tested to work fine.
* removes the setting to enable the pdns server, since all of the pdns
  servers are running versions that support it.
* updates the firewall rule to permit the pdns and designate components
  to use the cloud-private address only.
Also the configuration logic is centralized to the base profile from the
deployment-specific ones, and the types on the pdns_server class are
improved to catch some errors I did while developing this patch.

Bug: T336854
Bug: T348437
Change-Id: Ic9731f1456a6642ec5a41f9d22d5b32698a430c7