Publish metrics reflecting skein certificate expiry
The `skein.crt` certificates are currently auto-generated by airflow
itself, but we don't (yet) have eyes on the remaining days before these
certificates expire. This is a bit of a timebomb.

This patch aims to provide such metrics via a script exporting the
number of remaining days before expiration for these certificates, on
each airflow hosts.

Bug: T329398
Change-Id: Id6099a49829a1135059d54ae375c16e31e09df3f
Hosts: an-test-client1002.eqiad.wmnet
Hosts: an-launcher1002.eqiad.wmnet
Hosts: an-airflow1005.eqiad.wmnet