spicerack: enable distributed locking in production
Why:
In order to enable spicerack distributed locking support the etcd_config
configuration parameter should be set in spicerack's configuration file.

What:
Add an etcd_config parameter to the spicerack class and profile and set
the hiera value for it in the production and cloud related files.
Adapt the template that generates the configuration file to add the
etcd_config when present.

Hosts: auto
Bug: T341973
Change-Id: I2f805d6b183e59d984e365f064838e7aaa6bbf5a