Partial fix for multiple spark shufflers
The configuration for running multiple spark shufflers requires a
configuration directory and path for each of the spark-shuffle-site.xml
files.

Previously we were supplying this path by using the $HADOOP_CONF_DIR
variable, but this didn't work. It has been shown to work with an
absolute path without a variable.

However, the first spark shuffler service to get loaded doesn't honour
the custom configuration, so this is only a partial fix, which will get
the version 3.3 and 3.4 shufflers working.

Hosts: P:hadoop::common
Bug: T344910
Change-Id: I5c152b1e8fb714c3c38e9da5662535d9d198c372