Change the first spark shuffler service to use the default port
Our version of hadoop-yarn isn't loading the configuration overlay file
for the first version of the spark-yarn-shuffler service that gets
loaded.

On subsequent shuffler service loads it correctly picks up the custom
port number, but the first version to get loaded uses the default port
number of 7337.

This commit changes the spark3 default config to tell the client which
port to use for the default shuffler service.

Hosts: P:hadoop::common
Bug: T344910
Change-Id: Ie40af4b6bc1a686300012a5266c3ccbbd4cb81af