Fix issues with multiple spark shufflers specific to version 3.1
It was dicovered during testing that the option to use a named spark
shuffler service was only introduced in version 3.2 of spark.

Our current production pipelines are running on spark version 3.1.2 and
so we experienced errors attmpting to use the spark_shuffle_3_1 service.

This commit adds a number of conditionals, so that while we still use
version 3.1 it will call this version 'spark_shuffle` which is the
default.

Hosts: P:hadoop::common
Bug: T344910
Change-Id: I3ca7296525477d335e8768f4c81429e0b9ddf54d