httpd: ErrorLogFormat to strip fields with unavailable values
In Apache2 ErrorLogFormat, if one has:

  \"client.ip\": \"%a\", \"ecs.version\": "123"

When `%a` is not set, Apache will not emit the token and will not output
any non-whitespace surrounding the token. From the doc it "delete
everything from the preceding space character to the next space
character.". As a result we would end up with the invalid json:

            vvv
  "client.ip": "ecs.version": "123"
            ^^^

This change the JSONEncoder item separator to no more emit a whitespace,
Apache ErrorLogFormat thus considers the json key and value as a single
field and will strip them both when a placeholder does not exist.

If we require a field to be set, ErrorLogFormat supports setting a
default value when an item does not produce a value using `-` (minus).
For example `%-a` will ultimately result in:

  "client.ip":"-", "ecs.version":"123"

Although the behavior is only supported by `ErrorLogFormat`, regenerate
the existing `LogFormat` entries. The `generate_logformats.py` will gain
support for generating `ErrorLogFormat` lines in the follow-up change
I62bdaa430ba9fb63e4b0d4daefb38ffd489526b6.

Bug: T332672
Change-Id: I653e31f267bd088278fee9e39fa683ddb6c12bd9