Send metrics from Airflow analytics test
Configure the machine currently running the analytics_test instance
of Airflow to send a minimal list of metrics to Prometheus, using
statsd-exporter.

This patch was inspired by the existing monitoring of Superset.

The goal is a minimalist dashboard on Grafana where we can view
basic start, end, & duration data points.

At this point the new statsd-exporter endpoint will be included
in Prometheus-ops because of how prometheus::class_config is
working.

This patch focuses on the test cluster.

Bug: T349532
Bug: T343232
Change-Id: If53c120c183e52dbacf6914d6e9e7c5d586cc961