openstack: encapi: don't try to hold a single connection open
Unfortunately, the fix in 99e5a775953ac09782616497d04d14fdb82fac2e
didn't really work: the script no longer crashes, which is good, but now
it will not notice any changes made after the initial startup. That's
not good either.

So instead this commit changes the service to open a DB connection when
required, instead of trying to keep one open for a longer time. Given
that this service spends a vast majority of time executing a single
query, noticing it had zero results, and then waiting ten seconds before
retrying, that should be fine from a DB load perspective.

Bug: T349195
Change-Id: I2770883d70d986182feaa0cb31663bb6b6c32f6c