[systemd/timer] Add optional SuccessExitStatus argument to timer services
A common error that we see in rsync jobs is "some files vanished before
they could be transferred (code 24)", which happens when we try to rsync
a file that's been deleted after the job starts.

This error is one that we can ignore in many cases, but to do so within
rsync requires writing a wrapper script. So to solve it in a neater way,
we add the SuccessExitStatus flag in the .service file that is generated
for the timer.

Bug: T349166
Change-Id: I4f813dfb1baa2cdd1d459d566e7a50a635214ecb