Enable the management of the skein certificate via Puppet
If the hiera `profile::airflow::manage_skein_cerificate` is set to
`true`, we use our cfssl PKI to generate a certificate for skein, that
will be managed and renewed by Puppet runs.

For now, we don't enable this feature anywhere, so this should be a
no-op.

Bug: T329398
Hosts: an-airflow1002.eqiad.wmnet
Hosts: an-test-client1002.eqiad.wmnet
Change-Id: I99bdc2c9a36adf074de1a25e0cee830924af1ddd