mariadb: Remove sustained lag monitoring from misc dbs
Misc dbs, for several reasons, host applications that sometimes
produce noisy alerts due to spikes of lag. While we want to keep
monitoring large spikes of lag (regular monitoring based on
pt-heartbeat will keep being setup for large delays) in general,
sustained lag was supposed to just detect 1-2 seconds of lag for
a sustained time for mediawiki.

This should avoid a lot of noise on IRC- and will allow to focus
on real issues on mw dbs.

Change-Id: I6106cec7dacc7a1aafdbcf145ccca2cd8f36dc4c