piwik::dataase: Avoid Ferm-specific syntax
Remove Ferm-specific syntax from firewall definitions and switch to
firewall::service. This will eventually allow to migrate invididual
roles to nftables.

If passed an array of hosts, new srange() function revolves DNS names
on the puppetserver side (and that resolve happens transparently
for A/AAAA). As such modify and rename the parameter for the backup
DBs.

Hosts: matomo1002.eqiad.wmnet
Change-Id: I90f09b8431bdf05995d4fb0653ed4a421057fba2