puppet_compiler: always send CORS header even on 404
Old builds on https://puppet-compiler.wmflabs.org/ do not have JSON
representations of the build and requests are thus served as a 404.
Since Nginx `add_header` does not apply for 4xx or 5xx errors, there is
no `Access-Control-Allow-Origin` added and the browser bails out when
doing the CORS request.

Always add the CORS header and let the caller handle the 404.

Bug: T350003
Change-Id: I034b311fdd4eeae66b740e81b3d9ce84c7e6fd5f