switch druid host to run data_purge job
We are in the proces of onboarding new servers to the druid cluster and decommissioning
the older servers druid100[4-6]. As such, we shall need to change the druid-public endpoints used to
Drop Druid Public snapshots from deep storage following data retention policies.
It runs once a month for the datahub injestion jobs which is currently set to druid1004.eqiad.wmnet
we shall replace it with one of the newer servers namely druid1010.eqiad.wmnet.
druid1010 is already part of the running cluster thus we do not expect much interuption.

Bug: T336042
Change-Id: Id6c6a5020655a7613c1ace1253ac0eca32904f9e