prometheus: ipmi_exporter: add explicit dependency for sudo rule
Commit 8dca9efbb0 added an explicit depepdency to the IPMI exporter
which means that the service now successfully starts on the first Puppet
run. Unfortunately, it does that before the sudoers rule is provisioned,
resulting in rootspam. This commit fixes that by provisioning the sudo
rule before the package is even installed to ensure that it will be in
place by the time it's needed. (Apt starts the service when the package
is installed, so a dependency on the service unit is not quite enough.)

Change-Id: Ide56aa1024b407daa81525725418c346026b45cc