openstack: add wrapper for the common patch pattern
Adds a new resource to apply a patch to an OpenStack source file. The
main change in here is with ordering, now the dependency is applied to
the file too which means that Puppet no longer tries to add the patch
file to a directory that hasn't been created yet.

Also changes the file permissions from 0644 to 0444 like we usually do
for Puppet-managed files.

The client package directory patch is not included yet, it can be done
in a follow-up if needed.

Change-Id: Ice6b8c0e9e159ae7533699fcc2409e0b9b6bf231