Renew skein certificate every month via systemd timers
After having tried to issue a certificate via our cfssl PKI, and failing
to make it work with Skein, I'm now reverting to a simpler solution that
should still give us the level of reliability that we require.

We are now running a monthly command to renew the skein x509 certificate
(via the `skein` CLI) using a systemd timer.

Bug: T329398
Hosts: an-test-client1002.eqiad.wmnet
Change-Id: I4663551d99705f2b10982b8d5300fb0471b42a4d