openstack: replace openstack_controllers variable
This is a massive patch that replaces the current openstack_controllers
variable with a new one that has both the cloud-private and the main
host names included. The idea is that once all of the parents have been
merged, this one should be a PCC no-op (no resources changed on the
hosts), and then after this is merged we can go clean up the bits using
the wrong interfaces currently.

Bug: T347554
Change-Id: I8f1cf37011928e4ac0ea89738c534bc6c8c902dd