toolforge::docker::registry: update to use dnsquery::a
This Cr updates the code to use dnsquery::a instead of the older
ipresolve function. We also drop the explicit use of the nameservers
global variable. AFAIK the DNS servers of the compiling server will be
the same as the agent.

If not we can add the explicit nameserver back.

Bug: T350008
Hosts: P:toolforge::docker::registry
Change-Id: If1389f4fddfa867c9d6ae88ac2f02f64ee4a28ca