profile::openstack::base::puppetmaster::frontend: Avoid Ferm-specific syntax
These were added as Ferm rules, but can equally be represented as regular
service definitions.

While at it avoid Ferm-specific syntax and add these as a firewall::service.
This will eventually allow to migrate invididual roles to nftables.

(The srange() function revolves DNS names on the puppetserver side when passed
an array of hosts)

Change-Id: I233c214274a0bd8a82c9340c391a80cc4ca30aff