P:toolforge::mailrelay: rewrite maintainers in Python
Rewrites the maintainer lookup script from Bash to Python. This makes
implementing new features much easier than with the current shell
scripting monstrosity.

Also, for tools that do have maintainers but where none of the
maintainers have an email address set, mail is now sent into a black
hole instead of being deferred. This is meant to reduce the
constantly-large queue of deferred mail on the Toolforge mail servers.

The FIXME item for recursive tool maintainer processing is not yet
implemented.

Bug: T341006
Change-Id: I268de363be477e2b1381f60385fc6b7e1ec032e2