Use new mariadb server for analytics_meta
We have a new pair of MariaDB servers that are to be used for the
analytics_meta database that currently supports Hive, Druid, Superset,
DataHub, and Hue.

This commit updates all references to the Database on an-coord1001 to
an-mariadb1001 and should be merged at the same time as we promote the
database on an-mariadb1001 from a replica of an-coord1001 to be the new
master.

Hosts: C:bigtop::hue
Hosts: C:bigtop::hive
Hosts: C:druid
Hosts: C:superset
Bug: T284150
Change-Id: Idac974e7d38f6b1d6da34c8c49d3bd3655368198