mediawiki: Run expireTemporaryAccounts.php daily
For IP Masking, we want temporary accounts to expire
365 days afer they were first used. This is done by
a regularly executed maintenance script from CentralAuth.

To prevent "CentralAuth not found" error messages, the script
is executed on all wikis that are not private, fishbowl or non-global.
The script is also not executed on closed wikis, since those wikis
cannot be edited, and thus there is nothing to purge.

Hosts: mwmaint2002.codfw.wmnet, mwmaint1002.eqiad.wmnet
Bug: T344695
Change-Id: I8898daf1ca5a5c1a47a6d805e0c299684f1df3a5