hiera: re-order new cp hosts in eqiad
Previously, when we provisioned the new hosts in eqiad, we went with the
linear layout for the text/upload hosts as both of them share the same
configuration in eqiad, and as well as to keep it consistent with the
layout in other sites.

This results in the texts hosts being confined to rack A and B and
upload to rack C and D. For redundancy, we want the clusters to be
spread over the racks, so this commit picks the even numbers for text
and odd numbers for upload and re-orders the relevant configuration.

Bug: T349244
Change-Id: I87bef6356effe9fef0aea85ac0303d5f1d78a81a