Format both LVM volumes of an-druid1005 at next reimage
Due to an un-named feline troublemaker, the reimage process of
an-druid1005.eqiad.wmnet was disturbed, and as a result, the vg1-srv LVM
volume was _zeroed out_.

Any subsequent attempt to run the partman proess led to the `mk2fs`
process hanging for no apparent reason.

We're now in a state where  we'd rather just reformat everything from
scratch with the usual method.

Bug: T332604
Change-Id: I2a47663523fb1f1331c8350027d39f5ec623a939