Set an-master1003/1004 to use to Puppet 7 via Hiera host entries
These currently already use Puppet 7 by means of the
insetup::data_persistence role. However, since the current Hadoop master
and standby are still on Buster, the role isn't configured for Puppet 7.

As an interim measure add it via Hiera host entries, once
an-master1001/1002 are decommisioned, this can be re-applied to the
role instead.

Bug: T349619
Change-Id: I64761d237ee455df31485fe3766fe60183174b8f