Configure the analytics prometheus instance to start scraping airflow
We are starting to roll out statsd monitoring of airflow instances,
starting with the analytics_test instance. These will use
statsd_exporters that are deployed locally on each of the hosts hosting
an instance.

The intention is to use a separate prometheus job per instance, so that
this job label may be used to differentiate the same metrics between
different instances.

All of the current instances have been added in this commit, but only
the analytics_test instance is currently enabled for monitoring, so the
other six instances have been commented out for now.

Hosts: P:prometheus::ops
Hosts: P:prometheus::analytics
Bug: T343232
Change-Id: If6dc688a5d466fa23e197cc10dfa332b12c0422d