Add netboot config for new private vlans in codfw rows A/B
Had missed to add these when the vlans were originally defined and
added elsewhere.  I also changed the proxy url for current codfw
a/b vlans as it was set to eqiad.

We can merge this if needed or alternately if
I33663da17d0645e9f0d468f8e11dcb8e06bbb411 is ready just let that
take care of it.

Bug: T327938
Change-Id: Ie759fdc2e8758ae944bbc632d1dcf1e226665524