[toolsdb] Lower innodb_buffer_pool_size
This is not fixing the root cause of the OOM crashes we are seeing, but
it makes the OOMs less frequent while we find a proper fix.

I applied this change manually last Friday, but let's persist it in the
config file so it's not lost if MariaDB is restarted.

Bug: T349695
Change-Id: I9d8bdc7a945addfc3729e590fd18a39ff5d7c82a