profile::thanos: add new istio recording rule
The new recording rule is needed to support the lift wing pilot
for Pyrra, that IIUC already aggregates the raw data into recording
rules. The rule 'istio_sli_availability_requests_total' is needed
anyway to reduce the scope of the data pulled by Pyrra, but
in theory we could also bypass it. It seems more clear in this
way, but please let me know your thoughts.

Bug: T302995
Change-Id: Ifa2eee2d024dcff308966625c06855ef998b9585