puppetserver: cache code
Cache code forever, but trigger a cache eviction on every puppet-merge.
The aim of this change is to reduce race conditions when deploying code
via g10k as there is a period of time when the code directory is only
partially filled.

This change also manages /etc/puppet/puppetserver/conf.d recursively,
however the only change to the existing files is in auth.conf. Managing
all the files ensures there are less surprises during upgrades

Bug: T350809
Change-Id: I105eb177a88d8b45ac019c61666f6ae61c954f46