profile::pyrra::filesystem: improve/fix lift wing pilot
While checking the Pyrra dashboards I noticed some
numbers that didn't really make sense, like 50k rps
registered for some model servers. After a chat with
Filippo we suspect that the issue is related to the
new istio_sli_availability_requests recording rule,
that does a "sum" to aggregate some parameters,
not playing well with the rate() that Pyrra applies
to it.

This change moves the pilot to istio_requests_total,
the original metric before the recording rule.
There are a ton of metrics coming from Lift Wing,
but afaics Pyrra creates a recording rule called
something:increase12w (plus all the burn rate ones)
so in theory we should be good.

Bug: T302995
Change-Id: I59ca9694f92c73ceea7b6ba9c3a20d606d81d67c