Clean up hadoop coordinator roles by removing analytics_meta DB
We have now migrated the mariadb services from an-coord100[1-2] to
an-mariadb100[1-2] so we can remove the classes that added it from the
coordinator and cooridinator::replica roles.

This will not remove the packages nor data directory, but these servers
are to be refreshed shortly so that won't be required.

Hosts: an-coord1001.eqiad.wmnet,an-coord1002.eqiad.wmnet
Bug: T284150
Change-Id: I91f1866a3ba5991a55b744d52c7d9c1a9d65148f