Send recovery emails to data-engineering-alerts
The current configuration can be concerning for those engineers who
participate in the Data Engineering Ops Week rota because they receive
email when an alert is fired, but they do not receive an alert when the
alert recovers.

This can result in wasted time, as they investigate errors that were
transient or have been addressed already.

Hosts: C:alertmanager
Bug: T346438
Change-Id: I9d3d4a696938d5dc0eb9db0182cdff0afc5a835a