Configure Matomo's TagManager to write to existing tmpdir
We need to allow the TagManager plugin to write to a temporary directory
in order for it to function. By default, this would be
/usr/share/matomo/js but this directory currently does not have write
permission permitted.

Instead we wish to configure TagManager with a custom location for its
use, which we can do by providing a custom config.php file.

Hosts: O:piwik
Bug: T349910
Change-Id: Icf90c224160fbd002a32c2630630ef5243db4f53