conftool: introduce schema and host file for dnsboxes
Currently the list of active DNS servers is determined by the list in
hieradata/common.yaml under the authdns_server key. We use this list to
define the state of a DNS host -- whether it is pooled or not -- and the
maintenance of this list is done through Puppet changes which are
followed by agent runs on the affected hosts.

In an effort to reduce toil and as mentioned in the task below, we want
to move to a more dynamic state configuration, which we achieve by using
conftool. We introduce a new schema for the dnsbox role that takes into
account our requirements for this service.

Bug: T347054
Hosts: puppetmaster1001.eqiad.wmnet
Change-Id: Ided24097323a3e294fa9a765f59fa2a3fc470cb2