dbbackups: Update config for mysql backup monitoring
The backup configuration was pointing to the old, outdated
file, and while backups were running, they were failing to
report their status to the central db.

This patch will need a followup to cleanup the logic and
stop harcoding this variable, but leaving it like this
to unblock backups.

Document all the class arguments.

Bug: T351617
Bug: T351491
Change-Id: Ic4f95e9dc0aca75af7db29efa8683b872f6e6032