bacula: Increase the amount of maximum volumes for regular backups to 140
Currently, regular backups are limited to 120 volumes, or
120 * 500GB = 60TB of size, to prevent running out of space or
bad backups taking over all backup storage.

We are going to dedicate 10TB of additional space to regular backups
volume, for which we will allow 20 additional volumes.

This is ok, as eventually gerrit + gitlab will be split out in
a dedicated storage hosts that was purchased for it (probably
backup1010).

Bug: T351725
Change-Id: I612eeab448b3ab088f9ce314f8139c66edda8231