set druid hosts to use the reuse partman recipe
We are reimaging the druid100[7-8] instances to use bullseye,
this change sets them to use the reuse-raid10-4dev.cfg which tells partman to ignore
the raid array members, format the vg0-root partition to ext4 and remount it to / (the system) and
keep the vg0-srv LVM partition, mounted to /srv.
This shall allow a faster reimage to bullseye, same recipe was used for an-druid cluster.

Bug: T332589
Change-Id: Ide500478a272ebe51f7e3b70f2d1a8ade0926db8