airflow: change max_active_runs_per_dag back to 1
We recently increased the default value of max_active_runs_per_dag from
1 to 3 in order to facilitate faster backfill operations. However, after
we have now reviewed that decision in light of further experience and
have decided that default value of 1 is more sensible.

Users are still able to increase the value for individual jobs, but we
reduce the risk of accidental data corruption by resetting the default
back to 1.

Hosts: P:airflow
Bug: T351388
Bug: T300870
Change-Id: I5ab6ae7060108e90edaf0eb44c636d01415e6002