Fix Matomo TagManager functionality
The previous configuration attempt did not work because it did not
correctly interpret the wildcard in the path element of the Apache
virtualhost file.

This commit adds a new Require statement that tests whether the path is
exactly /js_container_*.js using a regex.

In addition to that, we were attempting to write the Javscript
containers to an existing directory with write access, but this didn't
work. In order to fix this we have removed the function override in
config.php and allowed write access to /usr/share/matomo/js for the
www-data user.

Hosts: O:piwik
Bug: T349910
Change-Id: I3ac2d8e729ea64dfa3d8f709021dcd28c7070efa