doc: move rsync auth secrets to new location to unbreak puppet
With change I13977a0d92cc599 in rsync::server the /etc/rsync.d/
directory is now removed.

This caused a puppet issue on doc hosts which are the only servers
that use this directory to place a secrets file into.

As suggested move the secrets file directly under /etc/ and update
the secrets_file param to unbreak this.

Bug: T351832
Change-Id: I43649819500e54b9bdee7d743b47bac01772ece3