puppetserver: use a symlink to swap in new code
Prior to this change we used g10k to deploy new code to the same
directory puppetserver was currently serving. Even with caching set to
unlimited, we still saw sporadic puppet failures during merges, as the
puppet code directory was in an inconsistent state. After this patch we
deploy to a staging directory and move a symlink, which should ensure
the code directory is internally consistent.

Bug: T350809
Hosts: C:puppetserver
Change-Id: I91a2395fbb01c33a556483a6923b510d3e6eae19