dbbackups: Execute the checks under the new Unix user backupcheck
Icinga needs to connect to the m1 dbbackups database to check
database backups metadata.

In order to avoid the mysql credentials to be publicly-readable,
make it owned by root and group backupcheck, the default group
of a newly created user backupcheck, which will be a system-only
user that will be able to read the mysql password.

The user matches the mysql user, like the generation uses 'dump'
for both the Unix and mysql user.

Change the commend to use absolute path to prevent sudo
escalation privileges.

Followup to 593519ff3e4de4ef707223ca0a9f5646886b3546

Change-Id: I388e70889812e2750484fc680d2016c11cf1820f
Hosts: backupmon1001.eqiad.wmnet