admin: reserve uid/gid for authdns user
Reserve the uid and gid for the authdns user on the DNS hosts. We
decided to do this so that we can pass these to confd::file, that
expects a uid and gid when creating the rendered template.

The current uid/gid is 495/1001 but we are reserving in preparation for
the next round of reimages where we will set it to 928/928.

Bug: T347054
Change-Id: I6953a12e80e19d4a49e3f0ea34821b900503e723