profile::pyrra::filesystem: reduce scope of the Lift Wing's pilot
The current approach doesn't work since Pyrra applied increase()
on a sum() aggregation, something that is discouraged by
upstream docs (sum of increases is fine, not the opposite). The effect
is the one stated in:

https://www.robustperception.io/rate-then-sum-never-sum-then-rate/

So the idea is to go back to the main istio_requests_total metric,
but to trim it down by service rather than using a big grouping
like the revscoring one.

In this case, the time series returned by the Thanos UI for error
are around 120, and for total ~400. This test should tell us if
Thanos rules can be managed if istio_requests_total is restricted
to the bare minimum, or if anything else is needed.

Bug: T351390
Change-Id: I2137449e7330598b9c2b18e61d290f56ee2d1313