P:dns::auth::update: add support for setting ferm rules via confd
This is the first commit that makes use of the dynamic configuration of
the DNS hosts to set the ferm rules, via confd. To do this, we template
the port 22 TCP rule for the DNS hosts that allows authdns-update to
work; this removes the dependence on the static authdns_servers key and
instead uses confd::file instead.

We make this a conditional on profile::dns::auth::confd_enabled so that
we can control how we roll this change out.

Bug: T347054
Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: I838a1624e7a7fb02ebd0e0e91e4fb713953d4763