P:dns::auth::update: add support for authdns-update hosts via confd
This commit is a continuation of the changes to the DNS hosts where the
configuration is set dynamically through confd. In this commit, we set
the list of active DNS hosts that is then used for the purpose of
authdns-udpate. If a host is not in this list, DNS changes are not
propagated to that host as part of running authdns-update.

This was being set through the authdns_servers key that necessitated a
Puppet change and an agent run. Doing this via confd will allow us to
push updates to all hosts without touching Puppet.

Bug: T347054
Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: I43c1077f94a7b085d585df67df8f3f54acccfe24