P:dns::auth::update: add support for generating .ssh/config via confd
As a replacement for the entries in /etc/hosts that include the IP
addresses for the DNS hosts so that they can reach other without a
working recursor (which is on the same host as the authdns ones), use an
.ssh/config file instead that is dynamically generated by confd.

Note that we don't care about the pooled status of a host here, simply
an active DNS host in the list, even if it is pooled or not. The pooled
state matters only in authdns-update itself to dictate which hosts
should get the updates.

This commit is intentionally separate from the chain in I838a1624e so as
to make it easier to merge these changes. We also remove the host {}
stanza that was managing these entries in /etc/hosts for us as that is
no longer required.

Bug: T347054
Hosts: dns1004.wikimedia.org, dns6001.wikimedia.org
Change-Id: Ifcf00272c69fd9e8b03345eff1c74d7599b2385a