hiera: dnsbox: remove anycast-hc dependency on pdns-rec
The current systemd dependency for the anycast-healthchecker.service
looks like:

    After=pdns-recursor.service
    BindsTo=pdns-recursor.service

Which means that if pdns-recursor is restarted or stopped, the
anycast-hc service is also restarted/stopped. This dependency then
trickles up to bird.service that depends on anycast-hc.service:

    Description=BIRD Internet Routing Daemon (BIRD2: IPv4 and IPv6)
    After=anycast-healthchecker.service
    BindsTo=anycast-healthchecker.service

This means that in case pdns-rec is stopped, bird is stopped, and that
withdraws the routes for all advertisements bird is doing. When it was
just pdns-recursor on the host announcing the recdns anycast address,
this was an acceptable design, but in T347054 we have moved the authdns
servers from their static routes to bird instead. (And also the NTP
service for the Debian installer.)

As a result of T347054, this means that if we stop pdns-recursor, bird
will also withdraw the advertisements for authdns/gdnsd as well. This is
not desirable and we would like to have more control over this; thus, we
are removing the pdns-rec dependency on anycast-hc, so that we can
withdraw pdns-rec routes individually and not affect the authdns routes.

Bug: T347054
Hosts: dns1004.wikimedia.org, dns6001.wikimedia.org, doh1001.wikimedia.org
Change-Id: Ie5c65799f4c14ccf342c8e0f6d2ab638f1a6a166